# refactor: simplify architecture by removing environment variables