… .tool-versions
… CLAUDE.md
… Cargo.lock
… Cargo.toml
… docs/PEFORMANCE_IMPROVEMENTS.md
… src/models/device_info.rs
… src/models/error.rs
… src/models/platform.rs
… src/ui/render.rs
… src/utils/command.rs
… src/utils/validation.rs
… tests/panel_switching_performance_test.rs
… tests/startup_performance_test.rs