## 📝 Description

This PR simplifies the architecture by removing all environment variable complexity and making performance optimizations always active by default. This eliminates the need for configuration and ensures consistent optimal performance for all users.

## 🧐 Type of change

- [x] ⚙️ Chore (refactoring, build process, CI, etc.)
- [x] ✨ New feature (performance improvements always enabled)
- [x] 📚 Documentation update

## 🔧 Changes Made

### Code Simplification
- Remove `use_*` environment check methods from App
- Always use optimized code paths:
  - Parallel commands execution
  - Fast panel switching
  - Smart device start
  - Incremental refresh
- Delete old non-parallel AndroidManager implementation
- Clean up dead code to resolve lint warnings

### Test Updates
- Remove environment variable setup from tests
- Update tests to reflect always-active optimizations
- All 108 tests pass

### Documentation Updates
- Update README.md to remove environment variable sections
- Update CLAUDE.md to reflect simplified architecture
- Document always-active performance features

## 🚀 Performance Improvements (Now Always Enabled)

- **Panel switching**: 30.8% faster (250ms → 173ms)
- **Device listing**: 42.2% faster (986ms → 571ms)
- **Incremental updates**: 10-40% faster
- **Device startup**: 200ms+ time savings

## ✅ How Has This Been Tested?

- [x] All 108 unit tests pass
- [x] Integration tests validate device lifecycle workflows
- [x] Performance tests confirm optimizations are active
- [x] Lint checks pass with no warnings
- [x] Manual testing of core functionality

**Test Configuration**:
- Platform: macOS (primary), with cross-platform compatibility
- Toolchain: Rust stable
- Dependencies: Android SDK, Xcode (for iOS)

## 🎯 Benefits

### For Users
- **Zero configuration**: Optimal performance out-of-the-box
- **Consistent experience**: All users get the same fast performance
- **Simplified usage**: No need to remember environment variables

### For Developers
- **Reduced complexity**: Simpler codebase without conditional paths
- **Lower maintenance**: Fewer potential configuration issues
- **Better reliability**: Less room for environment-related bugs

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added/updated tests that prove my changes work
- [x] New and existing unit tests pass locally with my changes
- [x] No dependent changes required

## 📊 Before vs After

### Before (with environment variables)
```bash
EMU_FAST_PANEL_SWITCH=true \
EMU_SMART_DEVICE_START=true \
EMU_INCREMENTAL_REFRESH=true \
EMU_PARALLEL_COMMANDS=true \
emu
```

### After (always optimized)
```bash
emu  # Maximum performance automatically
```

This change makes Emu truly "zero-config" while delivering the best possible performance to all users.